BE-613 Fix problems caused by backporting fixes for a newer Python

+++ src/PIL/GifImagePlugin.py
@@ -240,14 +240,13 @@ class GifImageFile(ImageFile.ImageFile):
                     self._size = max(x1, self.size[0]), max(y1, self.size[1])
                     Image._decompression_bomb_check(self._size)
                 self.dispose_extent = x0, y0, x1, y1
-                flags = s[8]
+                flags = i8(s[8])
 
                 interlace = (flags & 64) != 0
 
                 if flags & 128:
                     bits = (flags & 7) + 1
-                    self.palette = ImagePalette.raw(
-                        "RGB", self.fp.read(3 << bits))
+                    self.palette = ImagePalette.raw("RGB", self.fp.read(3 << bits))
 
                 # image data
                 bits = i8(self.fp.read(1))
@@ -268,8 +267,7 @@ class GifImageFile(ImageFile.ImageFile):
             elif self.disposal_method == 2:
                 # replace with background colour
                 Image._decompression_bomb_check(self.size)
-                self.dispose = Image.core.fill(
-                    "P", self.size, self.info["background"])
+                self.dispose = Image.core.fill("P", self.size, self.info["background"])
             else:
                 # replace with previous contents
                 if self.im:
@@ -307,8 +305,7 @@ class GifImageFile(ImageFile.ImageFile):
             # we do this by pasting the updated area onto the previous
             # frame which we then use as the current image content
             updated = self._crop(self.im, self.dispose_extent)
-            self._prev_im.paste(updated, self.dispose_extent,
-                                updated.convert("RGBA"))
+            self._prev_im.paste(updated, self.dispose_extent, updated.convert("RGBA"))
             self.im = self._prev_im
         self._prev_im = self.im.copy()
 
@@ -419,8 +416,7 @@ def _write_single_frame(im, fp, palette):
     _write_local_header(fp, im, (0, 0), flags)
 
     im_out.encoderconfig = (8, get_interlace(im))
-    ImageFile._save(
-        im_out, fp, [("gif", (0, 0) + im.size, 0, RAWMODE[im_out.mode])])
+    ImageFile._save(im_out, fp, [("gif", (0, 0) + im.size, 0, RAWMODE[im_out.mode])])
 
     fp.write(b"\0")  # end of image data
 
@@ -456,11 +452,9 @@ def _write_multiple_frames(im, fp, palette):
                     if background_im is None:
                         background = _get_background(
                             im,
-                            im.encoderinfo.get(
-                                "background", im.info.get("background")),
+                            im.encoderinfo.get("background", im.info.get("background")),
                         )
-                        background_im = Image.new(
-                            "P", im_frame.size, background)
+                        background_im = Image.new("P", im_frame.size, background)
                         background_im.putpalette(im_frames[0]["im"].palette)
                     base_im = background_im
                 else:
@@ -479,8 +473,7 @@ def _write_multiple_frames(im, fp, palette):
                     continue
             else:
                 bbox = None
-            im_frames.append({"im": im_frame, "bbox": bbox,
-                             "encoderinfo": encoderinfo})
+            im_frames.append({"im": im_frame, "bbox": bbox, "encoderinfo": encoderinfo})
 
     if len(im_frames) > 1:
         for frame_data in im_frames:
@@ -767,8 +760,7 @@ def _get_global_header(im, info):
     for extensionKey in ["transparency", "duration", "loop", "comment"]:
         if info and extensionKey in info:
             if (extensionKey == "duration" and info[extensionKey] == 0) or (
-                extensionKey == "comment" and not (
-                    1 <= len(info[extensionKey]) <= 255)
+                extensionKey == "comment" and not (1 <= len(info[extensionKey]) <= 255)
             ):
                 continue
             version = b"89a"
@@ -805,8 +797,7 @@ def _write_frame_data(fp, im_frame, offset, params):
         _write_local_header(fp, im_frame, offset, 0)
 
         ImageFile._save(
-            im_frame, fp, [("gif", (0, 0) + im_frame.size,
-                            0, RAWMODE[im_frame.mode])]
+            im_frame, fp, [("gif", (0, 0) + im_frame.size, 0, RAWMODE[im_frame.mode])]
         )
 
         fp.write(b"\0")  # end of image data
